+++ packages/core/src/index.ts
@@ -34,6 +34,7 @@ export {
   getLoadedLocales,
   getDictionary,
   getLoadedNamespaces,
+  hasNamespace,
   loadAsync,
   isLoaded,
   getLocaleDisplayName,